requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
asyncpg>=0.29
python-dotenv>=1.0
pydantic>=2.6
asyncpg
sqlalchemy[asyncio]
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
openai>=1.0.0
python-multipart
langgraph
langchain
langchain-google-genai
google-generativeai
python-dotenv
qdrant-client
qdrant-client
langchain-anthropic
langchain-community
pandas
openpyxl
nest_asyncio
google-auth
google-auth-oauthlib
google-api-python-client
python-telegram-bot
google-genai
elevenlabs
pg8000
psycopg2-binary
PyMuPDF
python-docx